+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/react
 
+## 0.48.0-pre.1
+
+### Patch Changes
+
+- @glissade/core@0.48.0-pre.1
+- @glissade/player@0.48.0-pre.1
+- @glissade/scene@0.48.0-pre.1
+
 ## 0.48.0-pre.0
 
 ### Patch Changes